Plutopay/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.2
django-cors-headers==4.3.1
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.12.2
urllib3==2.2.0
